+++ dev/user.clj
@@ -7,9 +7,12 @@
     [clj-java-decompiler.core :as decompiler]
     [clojure.tools.namespace.repl :as tns]
     [criterium.core :as criterium]
+    [taoensso.tufte :as tufte]
     [noahtheduke.splint.dev]))
 
 (defn refresh-all [& opts] (apply tns/refresh-all opts))
 (defmacro decompile [form] `(decompiler/decompile ~form))
 (defmacro quick-bench [expr & opts] `(criterium/quick-bench ~expr ~@opts))
-(defmacro bench [expr & opts] `(criterium/quick-bench ~expr ~@opts))
+(defmacro bench [expr & opts] `(criterium/bench ~expr ~@opts))
+(defmacro p [id & body] `(tufte/p ~id ~@body))
+(defmacro profile [opts & body] `(tufte/profile ~opts ~@body))