+++ test/noahtheduke/splint/runner_test.clj
@@ -4,6 +4,7 @@
 
 (ns noahtheduke.splint.runner-test
   (:require
+    [clojure.java.io :as io]
     [expectations.clojure.test :refer [defexpect]]
     [noahtheduke.splint.test-helpers :refer [expect-match]]))
 
@@ -20,3 +21,39 @@
   (expect-match
     '[{:rule-name style/plus-one}]
     "#_{:splint/disable [style/plus-zero]} (+ 1 x)"))
+
+(defexpect throws-test
+  (expect-match
+    '[{:rule-name splint/error
+       :form (very-special-symbol :do-not-match)
+       :message "Splint encountered an error: \"matched\""}]
+    "(very-special-symbol :do-not-match)")
+  (expect-match
+    '[{:rule-name naming/single-segment-namespace
+       :form (ns throw-in-middle)
+       :message "throw-in-middle is a single segment. Consider adding an additional segment."
+       :alt nil
+       :line 5
+       :column 1
+       :end-row 5
+       :end-col 21
+       :filename "corpus/throw_in_middle.clj"}
+      {:rule-name splint/error
+       :form (very-special-symbol :do-not-match)
+       :message "Splint encountered an error: \"matched\""
+       :alt nil
+       :line 7
+       :column 1
+       :end-row 7
+       :end-col 36
+       :filename "corpus/throw_in_middle.clj"}
+      {:rule-name lint/let-if
+       :form (let [a 1] (if a (+ a a) 2))
+       :message "Use `if-let` instead of recreating it."
+       :alt (if-let [a 1] (+ a a) 2)
+       :line 9
+       :column 1
+       :end-col 29
+       :end-row 9
+       :filename "corpus/throw_in_middle.clj"}]
+    (io/file "corpus" "throw_in_middle.clj")))